create dap functions module
