+++ crates/dapf/src/functions/mod.rs
@@ -1,0 +1,6 @@
+//! The various DAP functions dapf can perform.
+// Copyright (c) 2024 Cloudflare, Inc. All rights reserved.
+// SPDX-License-Identifier: BSD-3-Clause
+
+pub mod hpke;
+pub mod test_routes;
